Do basic, introductory drawing with SDL2 library.
--------- Key points related to this commit -------------
In this commit, the first steps in working with SDL2 are made.
The equivalent to the standard "Hello, World!" beginners task is done, but it is done in a graphical manner with SDL2.

Examples of basic SDL2 functionalities that are used:
Initialisation / Deinitialisation of SDL2,
Creating SDL2 surfaces, creating a GUI window,
Loading image graphical resources from the HDD into CPU primitives (i.e. SDL_Surfaces)
Rendering the CPU primitives on the screen is performed
Destroying SDL_Surfaces and the GUI window
Some more ...

The above functionalities will be largely utilised and enhanced in the following commits.
Very, very basic code architecture concepts are presented as well in the main.cpp file.

----- Brief explanation of major classes / objects -------

0. cmake_helpers folder
- contains a "find_modules" folder which, in brief, enables CMake to find the location of the SDL libraries on our machines
- contains a "helpers.cmake" file which offers two auxiliary CMake functions:
1) The first function sets specifics related to the language standard, extentions, etc... which CMake will consider
2) The second function sets the type of warnings which the compiler must return, if encountered in the code

1. main.cpp file
- for the current moment, contains the entire program code /application

The building of the basis of the chess game application
is NOT started yet.