Do basic, introductory drawing with SDL2 library.
In this commit, the first steps in working with SDL2
are made. The equivalent to the standard "Hello, World!" beginners task
is done, but it is done in a graphical manner with SDL2.

Usage of basic SDL2 functions:
Initialisation / Deinitialisation of SDL2,
Creating SDL2 surfaces / windows,
Loading graphical information from the HDD and drawing it on SDL2
surfaces, Destroying SDL2 surfaces/windows,
Some more ...

Very basic software architecture concepts are presented as well.

The building of the basis of the chess game application
is NOT started yet.